moved all the files into distances
