+++ sktime/distances/__init__.py
@@ -1,0 +1,32 @@
+# -*- coding: utf-8 -*-
+"""Distance computation."""
+
+__all__ = [
+    "distance",
+    "distance_factory",
+    "pairwise_distance",
+    "euclidean_distance",
+    "squared_distance",
+    "dtw_distance",
+    "ddtw_distance",
+    "wdtw_distance",
+    "wddtw_distance",
+    "edr_distance",
+    "lcss_distance",
+    "LowerBounding",
+]
+
+from sktime.distances.distance import (
+    ddtw_distance,
+    distance,
+    distance_factory,
+    dtw_distance,
+    edr_distance,
+    euclidean_distance,
+    lcss_distance,
+    pairwise_distance,
+    squared_distance,
+    wddtw_distance,
+    wdtw_distance,
+)
+from sktime.distances.lower_bounding import LowerBounding
