+++ sktime/distances/_dtw.py
@@ -6,17 +6,14 @@ from typing import Callable, Union
 import numpy as np
 from numba import njit
 
-from sktime.dists_kernels.numba.distances._numba_utils import _compute_pairwise_distance
-from sktime.dists_kernels.numba.distances._squared_distance import _SquaredDistance
-from sktime.dists_kernels.numba.distances.base import DistanceCallable, NumbaDistance
-from sktime.dists_kernels.numba.distances.dtw_based.lower_bounding import (
-    LowerBounding,
-    resolve_bounding_matrix,
-)
+from sktime.distances._numba_utils import _compute_pairwise_distance
+from sktime.distances._squared import _SquaredDistance
+from sktime.distances.base import DistanceCallable, NumbaDistance
+from sktime.distances.lower_bounding import LowerBounding, resolve_bounding_matrix
 
 
-class _EdrDistance(NumbaDistance):
-    """Edit distance for real sequences (edr) between two timeseries."""
+class _DtwDistance(NumbaDistance):
+    """Dynamic time warping (dtw) between two timeseries."""
 
     def _distance_factory(
         self,
@@ -27,10 +24,9 @@ class _EdrDistance(NumbaDistance):
         itakura_max_slope: float = 2.0,
         custom_distance: DistanceCallable = _SquaredDistance().distance_factory,
         bounding_matrix: np.ndarray = None,
-        epsilon: float = 0.5,
         **kwargs: dict
     ) -> DistanceCallable:
-        """Create a no_python compiled edr distance callable.
+        """Create a no_python compiled dtw distance callable.
 
         Parameters
         ----------
@@ -54,12 +50,10 @@ class _EdrDistance(NumbaDistance):
             and creation are ignored. The matrix should be structure so that indexes
             considered in bound should be the value 0. and indexes outside the bounding
             matrix should be infinity.
-        epsilon : float, defaults = 0.5
-            Matching threshold to determine if two subsequences are considered close
-            enough to be considered 'common'.
         kwargs: dict
             Extra arguments for custom distance should be put in the kwargs. See the
             documentation for the distance for kwargs.
+
 
         Returns
         -------
@@ -80,31 +74,28 @@ class _EdrDistance(NumbaDistance):
 
         # This needs to be here as potential distances only known at runtime not
         # compile time so having this at the top would cause circular import errors.
-        from sktime.dists_kernels.numba.distances.distance import distance_factory
+        from sktime.distances.distance import distance_factory
 
         _custom_distance = distance_factory(x, y, metric=custom_distance, **kwargs)
 
         @njit()
-        def numba_edr_distance(
+        def numba_dtw_distance(
             _x: np.ndarray,
             _y: np.ndarray,
         ) -> float:
-            return _numba_edr_distance(
-                _x, _y, _custom_distance, _bounding_matrix, epsilon
-            )
+            return _dtw_numba_distance(_x, _y, _custom_distance, _bounding_matrix)
 
-        return numba_edr_distance
+        return numba_dtw_distance
 
 
 @njit()
-def _numba_edr_distance(
+def _dtw_numba_distance(
     x: np.ndarray,
     y: np.ndarray,
-    distance: Callable[[np.ndarray, np.ndarray], float],
+    custom_distance: Callable[[np.ndarray, np.ndarray], float],
     bounding_matrix: np.ndarray,
-    epsilon: float,
 ) -> float:
-    """Edr distance compiled to no_python.
+    """Dtw distance compiled to no_python.
 
     Parameters
     ----------
@@ -112,40 +103,34 @@ def _numba_edr_distance(
         First timeseries.
     y: np.ndarray (2d array)
         Second timeseries.
-    distance: Callable[[np.ndarray, np.ndarray], float],
-                    defaults = squared_distance
+    custom_distance: Callable[[np.ndarray, np.ndarray], float],
         Distance function to used to compute distance between timeseries.
     bounding_matrix: np.ndarray (2d of size mxn where m is len(x) and n is len(y))
         Bounding matrix where the values in bound are marked by finite values and
         outside bound points are infinite values.
-    epsilon : float
-        Matching threshold to determine if two subsequences are considered close enough
-        to be considered 'common'.
 
     Returns
     -------
-    float
-        Edr between two timeseries.
+    distance: float
+        Dtw distance between the two timeseries.
     """
     symmetric = np.array_equal(x, y)
-    pre_computed_distances = _compute_pairwise_distance(x, y, symmetric, distance)
-
-    cost_matrix = _edr_cost_matrix(
-        x, y, bounding_matrix, pre_computed_distances, epsilon
+    pre_computed_distances = _compute_pairwise_distance(
+        x, y, symmetric, custom_distance
     )
 
-    return float(cost_matrix[-1, -1] / max(x.shape[0], y.shape[0]))
+    cost_matrix = _cost_matrix(x, y, bounding_matrix, pre_computed_distances)
+    return cost_matrix[-1, -1]
 
 
-@njit()
-def _edr_cost_matrix(
+@njit(cache=True)
+def _cost_matrix(
     x: np.ndarray,
     y: np.ndarray,
     bounding_matrix: np.ndarray,
     pre_computed_distances: np.ndarray,
-    epsilon: float,
 ):
-    """Compute the edr cost matrix between two timeseries.
+    """Compute the dtw cost matrix between two timeseries.
 
     Parameters
     ----------
@@ -158,30 +143,23 @@ def _edr_cost_matrix(
         outside bound points are infinite values.
     pre_computed_distances: np.ndarray (2d of size mxn where m is len(x) and n is
                                         len(y))
-        Pre-computed distances.
-    epsilon : float
-        Matching threshold to determine if two subsequences are considered close enough
-        to be considered 'common'.
+        Precomputed pairwise matrix between the two timeseries.
 
     Returns
     -------
-    np.ndarray (2d of size mxn where m is len(x) and n is len(y))
-        Edr cost matrix between x and y.
+    np.ndarray (2d array of size mxn where m is len(x) and n is len(y))
+        Cost matrix between two timeseries.
     """
     x_size = x.shape[0]
     y_size = y.shape[0]
-    cost_matrix = np.zeros((x_size + 1, y_size + 1))
+    cost_matrix = np.full((x_size + 1, y_size + 1), np.inf)
+    cost_matrix[0, 0] = 0.0
 
-    for i in range(1, x_size + 1):
-        for j in range(1, y_size + 1):
-            if np.isfinite(bounding_matrix[i - 1, j - 1]):
-                if pre_computed_distances[i - 1, j - 1] < epsilon:
-                    cost = 0
-                else:
-                    cost = 1
-                cost_matrix[i, j] = min(
-                    cost_matrix[i - 1, j - 1] + cost,
-                    cost_matrix[i - 1, j] + 1,
-                    cost_matrix[i, j - 1] + 1,
+    for i in range(x_size):
+        for j in range(y_size):
+            if np.isfinite(bounding_matrix[i, j]):
+                cost_matrix[i + 1, j + 1] = pre_computed_distances[i, j] + min(
+                    cost_matrix[i, j + 1], cost_matrix[i + 1, j], cost_matrix[i, j]
                 )
+
     return cost_matrix[1:, 1:]
