+++ sktime/distances/_edr.py
@@ -6,17 +6,14 @@ from typing import Callable, Union
 import numpy as np
 from numba import njit
 
-from sktime.dists_kernels.numba.distances._numba_utils import _compute_pairwise_distance
-from sktime.dists_kernels.numba.distances._squared_distance import _SquaredDistance
-from sktime.dists_kernels.numba.distances.base import DistanceCallable, NumbaDistance
-from sktime.dists_kernels.numba.distances.dtw_based.lower_bounding import (
-    LowerBounding,
-    resolve_bounding_matrix,
-)
+from sktime.distances._numba_utils import _compute_pairwise_distance
+from sktime.distances._squared import _SquaredDistance
+from sktime.distances.base import DistanceCallable, NumbaDistance
+from sktime.distances.lower_bounding import LowerBounding, resolve_bounding_matrix
 
 
-class _DtwDistance(NumbaDistance):
-    """Dynamic time warping (dtw) between two timeseries."""
+class _EdrDistance(NumbaDistance):
+    """Edit distance for real sequences (edr) between two timeseries."""
 
     def _distance_factory(
         self,
@@ -27,9 +24,10 @@ class _DtwDistance(NumbaDistance):
         itakura_max_slope: float = 2.0,
         custom_distance: DistanceCallable = _SquaredDistance().distance_factory,
         bounding_matrix: np.ndarray = None,
+        epsilon: float = 0.5,
         **kwargs: dict
     ) -> DistanceCallable:
-        """Create a no_python compiled dtw distance callable.
+        """Create a no_python compiled edr distance callable.
 
         Parameters
         ----------
@@ -53,10 +51,12 @@ class _DtwDistance(NumbaDistance):
             and creation are ignored. The matrix should be structure so that indexes
             considered in bound should be the value 0. and indexes outside the bounding
             matrix should be infinity.
+        epsilon : float, defaults = 0.5
+            Matching threshold to determine if two subsequences are considered close
+            enough to be considered 'common'.
         kwargs: dict
             Extra arguments for custom distance should be put in the kwargs. See the
             documentation for the distance for kwargs.
-
 
         Returns
         -------
@@ -77,28 +77,31 @@ class _DtwDistance(NumbaDistance):
 
         # This needs to be here as potential distances only known at runtime not
         # compile time so having this at the top would cause circular import errors.
-        from sktime.dists_kernels.numba.distances.distance import distance_factory
+        from sktime.distances.distance import distance_factory
 
         _custom_distance = distance_factory(x, y, metric=custom_distance, **kwargs)
 
         @njit()
-        def numba_dtw_distance(
+        def numba_edr_distance(
             _x: np.ndarray,
             _y: np.ndarray,
         ) -> float:
-            return _dtw_numba_distance(_x, _y, _custom_distance, _bounding_matrix)
+            return _numba_edr_distance(
+                _x, _y, _custom_distance, _bounding_matrix, epsilon
+            )
 
-        return numba_dtw_distance
+        return numba_edr_distance
 
 
-@njit(cache=True)
-def _dtw_numba_distance(
+@njit()
+def _numba_edr_distance(
     x: np.ndarray,
     y: np.ndarray,
-    custom_distance: Callable[[np.ndarray, np.ndarray], float],
+    distance: Callable[[np.ndarray, np.ndarray], float],
     bounding_matrix: np.ndarray,
+    epsilon: float,
 ) -> float:
-    """Dtw distance compiled to no_python.
+    """Edr distance compiled to no_python.
 
     Parameters
     ----------
@@ -106,34 +109,40 @@ def _dtw_numba_distance(
         First timeseries.
     y: np.ndarray (2d array)
         Second timeseries.
-    custom_distance: Callable[[np.ndarray, np.ndarray], float],
+    distance: Callable[[np.ndarray, np.ndarray], float],
+                    defaults = squared_distance
         Distance function to used to compute distance between timeseries.
     bounding_matrix: np.ndarray (2d of size mxn where m is len(x) and n is len(y))
         Bounding matrix where the values in bound are marked by finite values and
         outside bound points are infinite values.
+    epsilon : float
+        Matching threshold to determine if two subsequences are considered close enough
+        to be considered 'common'.
 
     Returns
     -------
-    distance: float
-        Dtw distance between the two timeseries.
+    float
+        Edr between two timeseries.
     """
     symmetric = np.array_equal(x, y)
-    pre_computed_distances = _compute_pairwise_distance(
-        x, y, symmetric, custom_distance
+    pre_computed_distances = _compute_pairwise_distance(x, y, symmetric, distance)
+
+    cost_matrix = _edr_cost_matrix(
+        x, y, bounding_matrix, pre_computed_distances, epsilon
     )
 
-    cost_matrix = _cost_matrix(x, y, bounding_matrix, pre_computed_distances)
-    return cost_matrix[-1, -1]
+    return float(cost_matrix[-1, -1] / max(x.shape[0], y.shape[0]))
 
 
-@njit(cache=True)
-def _cost_matrix(
+@njit()
+def _edr_cost_matrix(
     x: np.ndarray,
     y: np.ndarray,
     bounding_matrix: np.ndarray,
     pre_computed_distances: np.ndarray,
+    epsilon: float,
 ):
-    """Compute the dtw cost matrix between two timeseries.
+    """Compute the edr cost matrix between two timeseries.
 
     Parameters
     ----------
@@ -146,23 +155,30 @@ def _cost_matrix(
         outside bound points are infinite values.
     pre_computed_distances: np.ndarray (2d of size mxn where m is len(x) and n is
                                         len(y))
-        Precomputed pairwise matrix between the two timeseries.
+        Pre-computed distances.
+    epsilon : float
+        Matching threshold to determine if two subsequences are considered close enough
+        to be considered 'common'.
 
     Returns
     -------
-    np.ndarray (2d array of size mxn where m is len(x) and n is len(y))
-        Cost matrix between two timeseries.
+    np.ndarray (2d of size mxn where m is len(x) and n is len(y))
+        Edr cost matrix between x and y.
     """
     x_size = x.shape[0]
     y_size = y.shape[0]
-    cost_matrix = np.full((x_size + 1, y_size + 1), np.inf)
-    cost_matrix[0, 0] = 0.0
+    cost_matrix = np.zeros((x_size + 1, y_size + 1))
 
-    for i in range(x_size):
-        for j in range(y_size):
-            if np.isfinite(bounding_matrix[i, j]):
-                cost_matrix[i + 1, j + 1] = pre_computed_distances[i, j] + min(
-                    cost_matrix[i, j + 1], cost_matrix[i + 1, j], cost_matrix[i, j]
+    for i in range(1, x_size + 1):
+        for j in range(1, y_size + 1):
+            if np.isfinite(bounding_matrix[i - 1, j - 1]):
+                if pre_computed_distances[i - 1, j - 1] < epsilon:
+                    cost = 0
+                else:
+                    cost = 1
+                cost_matrix[i, j] = min(
+                    cost_matrix[i - 1, j - 1] + cost,
+                    cost_matrix[i - 1, j] + 1,
+                    cost_matrix[i, j - 1] + 1,
                 )
-
     return cost_matrix[1:, 1:]
