+++ sktime/distances/_lcss.py
@@ -4,17 +4,14 @@ from typing import Callable, Union
 import numpy as np
 from numba import njit
 
-from sktime.dists_kernels.numba.distances._numba_utils import _compute_pairwise_distance
-from sktime.dists_kernels.numba.distances._squared_distance import _SquaredDistance
-from sktime.dists_kernels.numba.distances.base import DistanceCallable, NumbaDistance
-from sktime.dists_kernels.numba.distances.dtw_based.lower_bounding import (
-    LowerBounding,
-    resolve_bounding_matrix,
-)
+from sktime.distances._numba_utils import _compute_pairwise_distance
+from sktime.distances._squared import _SquaredDistance
+from sktime.distances.base import DistanceCallable, NumbaDistance
+from sktime.distances.lower_bounding import LowerBounding, resolve_bounding_matrix
 
 
-class _WdtwDistance(NumbaDistance):
-    """Weighted dynamic time warping (Wdtw) distance between two timeseries."""
+class _LcssDistance(NumbaDistance):
+    """Longest common subsequence (Lcss) between two timeseries."""
 
     def _distance_factory(
         self,
@@ -25,10 +22,10 @@ class _WdtwDistance(NumbaDistance):
         itakura_max_slope: float = 2.0,
         custom_distance: DistanceCallable = _SquaredDistance().distance_factory,
         bounding_matrix: np.ndarray = None,
-        g: float = 0.0,
+        epsilon: float = 1.0,
         **kwargs: dict,
     ) -> DistanceCallable:
-        """Create a no_python compiled wdtw distance callable.
+        """Create a no_python compiled lcss distance callable.
 
         Parameters
         ----------
@@ -52,10 +49,9 @@ class _WdtwDistance(NumbaDistance):
             and creation are ignored. The matrix should be structure so that indexes
             considered in bound should be the value 0. and indexes outside the bounding
             matrix should be infinity.
-        g: float, defaults = 0.
-            Constant that controls the curvature (slope) of the function; that is, g
-            controls the level of penalisation for the points with larger phase
-            difference.
+        epsilon : float, defaults = 1.
+            Matching threshold to determine if two subsequences are considered close
+            enough to be considered 'common'.
         kwargs: dict
             Extra arguments for custom distance should be put in the kwargs. See the
             documentation for the distance for kwargs.
@@ -63,7 +59,7 @@ class _WdtwDistance(NumbaDistance):
         Returns
         -------
         Callable[[np.ndarray, np.ndarray], float]
-            No_python compiled wdtw distance callable.
+            No_python compiled lcss distance callable.
 
         Raises
         ------
@@ -79,29 +75,31 @@ class _WdtwDistance(NumbaDistance):
 
         # This needs to be here as potential distances only known at runtime not
         # compile time so having this at the top would cause circular import errors.
-        from sktime.dists_kernels.numba.distances.distance import distance_factory
+        from sktime.distances.distance import distance_factory
 
         _custom_distance = distance_factory(x, y, metric=custom_distance, **kwargs)
 
         @njit()
-        def numba_wdtw_distance(
+        def numba_lcss_distance(
             _x: np.ndarray,
             _y: np.ndarray,
         ) -> float:
-            return _wdtw_numba_distance(_x, _y, _custom_distance, _bounding_matrix, g)
+            return _numba_lcss_distance(
+                _x, _y, _custom_distance, _bounding_matrix, epsilon
+            )
 
-        return numba_wdtw_distance
+        return numba_lcss_distance
 
 
 @njit(cache=True)
-def _wdtw_numba_distance(
+def _numba_lcss_distance(
     x: np.ndarray,
     y: np.ndarray,
-    custom_distance: Callable[[np.ndarray, np.ndarray], float],
+    distance: Callable[[np.ndarray, np.ndarray], float],
     bounding_matrix: np.ndarray,
-    g: float,
+    epsilon: float,
 ) -> float:
-    """Wdtw distance compiled to no_python.
+    """Lcss distance compiled to no_python.
 
     Parameters
     ----------
@@ -109,40 +107,39 @@ def _wdtw_numba_distance(
         First timeseries.
     y: np.ndarray (2d array)
         Second timeseries.
-    custom_distance: Callable[[np.ndarray, np.ndarray], float],
+    distance: Callable[[np.ndarray, np.ndarray], float],
+                    defaults = squared_distance
         Distance function to used to compute distance between timeseries.
     bounding_matrix: np.ndarray (2d of size mxn where m is len(x) and n is len(y))
         Bounding matrix where the values in bound are marked by finite values and
         outside bound points are infinite values.
-    g: float
-        Constant that controls the curvature (slope) of the function; that is, g
-        controls the level of penalisation for the points with larger phase difference.
+    epsilon : float
+        Matching threshold to determine if two subsequences are considered close enough
+        to be considered 'common'.
 
     Returns
     -------
-    distance: float
-        Wdtw distance between the two timeseries.
+    float
+        lcss between two timeseries.
     """
     symmetric = np.array_equal(x, y)
-    pre_computed_distances = _compute_pairwise_distance(
-        x, y, symmetric, custom_distance
-    )
+    pre_computed_distances = _compute_pairwise_distance(x, y, symmetric, distance)
 
-    cost_matrix = _weighted_cost_matrix(
-        x, y, bounding_matrix, pre_computed_distances, g
+    cost_matrix = _sequence_cost_matrix(
+        x, y, bounding_matrix, pre_computed_distances, epsilon
     )
-    return np.sqrt(cost_matrix[-1, -1])
+    return float(cost_matrix[-1, -1] / min(x.shape[0], y.shape[0]))
 
 
 @njit(cache=True)
-def _weighted_cost_matrix(
+def _sequence_cost_matrix(
     x: np.ndarray,
     y: np.ndarray,
     bounding_matrix: np.ndarray,
     pre_computed_distances: np.ndarray,
-    g: float,
+    epsilon: float,
 ):
-    """Compute the wdtw cost matrix between two timeseries.
+    """Compute the lcss cost matrix between two timeseries.
 
     Parameters
     ----------
@@ -155,31 +152,28 @@ def _weighted_cost_matrix(
         outside bound points are infinite values.
     pre_computed_distances: np.ndarray (2d of size mxn where m is len(x) and n is
                                         len(y))
-        Precomputed pairwise matrix between the two timeseries.
-    g: float
-        Constant that controls the curvature (slope) of the function; that is, g
-        controls the level of penalisation for the points with larger phase difference.
+        Pre-computed distances.
+    epsilon : float
+        Matching threshold to determine if two subsequences are considered close enough
+        to be considered 'common'.
 
     Returns
     -------
-    np.ndarray
-        Weighted cost matrix between x and y timeseries
+    np.ndarray (2d of size mxn where m is len(x) and n is len(y))
+        Lcss cost matrix between x and y.
     """
     x_size = x.shape[0]
     y_size = y.shape[0]
-    cost_matrix = np.full((x_size + 1, y_size + 1), np.inf)
-    cost_matrix[0, 0] = 0.0
+    cost_matrix = np.zeros((x_size + 1, y_size + 1))
 
-    weight_vector = np.array(
-        [1 / (1 + np.exp(-g * (i - x_size / 2))) for i in range(0, x_size)]
-    )
+    for i in range(1, x_size + 1):
+        for j in range(1, y_size + 1):
+            if np.isfinite(bounding_matrix[i - 1, j - 1]):
+                if pre_computed_distances[i - 1, j - 1] <= epsilon:
+                    cost_matrix[i, j] = 1 + cost_matrix[i - 1, j - 1]
+                else:
+                    cost_matrix[i, j] = max(
+                        cost_matrix[i, j - 1], cost_matrix[i - 1, j]
+                    )
 
-    for i in range(x_size):
-        for j in range(y_size):
-            if np.isfinite(bounding_matrix[i, j]):
-                cost_matrix[i + 1, j + 1] = (
-                    min(cost_matrix[i, j + 1], cost_matrix[i + 1, j], cost_matrix[i, j])
-                    + weight_vector[np.abs(i - j)] * pre_computed_distances[i, j]
-                )
-
-    return cost_matrix[1:, 1:]
+    return cost_matrix
