--- sktime/dists_kernels/numba/distances/__init__.py
@@ -1,22 +1,0 @@
-# -*- coding: utf-8 -*-
-"""Module containing numba compiled distances and utilities."""
-
-__author__ = ["chrisholder"]
-__all__ = [
-    "pairwise_distance",
-    "euclidean_distance",
-    "pairwise_euclidean_distance",
-    "squared_distance",
-    "pairwise_squared_distance",
-]
-
-
-from sktime.dists_kernels.numba.distances.euclidean_distance import (
-    euclidean_distance,
-    pairwise_euclidean_distance,
-)
-from sktime.dists_kernels.numba.distances.pairwise_distance import pairwise_distance
-from sktime.dists_kernels.numba.distances.squared_distance import (
-    pairwise_squared_distance,
-    squared_distance,
-)
